Update 1.setup.md
Hi.
Reviewing docs:  I read it as a setup for Nuxt 3 Tailwind, however the nuxt.config refers to the old 'export default' vs. the new direction.  Thus, suggesting stacking code boxes, with reference to the versions to load, as a possible solution?

ALSO... on the 'home page' for this site, it has an impage of the terminal with instructions.  That is great, however I'd also caption that image with 'See here for next steps', with hyper link to the next page.  I missread the home page last time I visited this site as 'that is all I needed to do'... I am very much a bigginner, thus the mistake.

Love your work.  You guys rock.

Thanks.  Nick